fix : 나머지 오류 수정
