+++ src/service/client-actions/portfolio.ts
@@ -46,33 +46,6 @@ export const postPortfolio = async (data: any) => {
   return false;
 };
 
-export const getFeeds = async ({
-  pageParam = 0,
-  role = 'all',
-}: {
-  pageParam: number;
-  role: string;
-}) => {
-  const rolePathMap: Record<string, string> = {
-    all: '',
-    model: '/model',
-    author: '/photography',
-  };
-
-  const path = rolePathMap[role] ?? '';
-  const url = `${API_URL}/portfolios${path}?page=${pageParam}&size=10`;
-
-  const res = await fetch(url, {
-    method: 'GET',
-    cache: 'no-store',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  });
-
-  return res.json();
-};
-
 export const getPortfolioDetailClient = async (id?: string) => {
   const token = getCookie('accessToken');
